[PI-19] Refactor crypto methods and clean-up in WalletStorage.js
